Entrypoint.sh: Start xdebug on request removed
You can start the debugging with an .env variable or by a cookie setting, see https://xdebug.org/docs/all_settings#start_with_request. Without starting the debugging on  request by default, the image is also usable in prod environment or at least you have the choice if you want debugging in an environment or not.